requirements.txt
pillow==9.2.0
tk==8.6.12
